Replace sphinx manpage with manual one.
If we upload source tarballs to PyPI instead of eggs, then the man
pages actually work. However, this means that the installer has to
run setup.py. Using sphinx, while slighly nicer from a dev
perspective, means we have to then require that people installing
git-review also install sphinx, which is a bit overkill given the
size of the man page.

This change, and a change to the release process of not uploading
eggs to PyPI should finall get git review --help to work.

Change-Id: I29b07667127a9ec5c241b7e1fc5a3960dd2a6b63