Add --finish option to git-review.
Closes bug 902858.

Folks using git flow are used to using a command, "git flow finish" which
merges their changes back in to develop and then deletes the branch. We
obviously do not want our git flow users to do that, since we don't want the
merge commits. So this adds a -f/--finish option which allows you to get rid
of the topic branch after successful submission.

Change-Id: I0f3b2a5b4826de074480b684ec9b5603788f15b7