Remove commit --amend from hook installation.
Fixes bug 911362.

If you run git review -s in a fresh repo, there is no reason for us to
amend any commits to add a commit message.

If you run git review in a repo with a change but which has not been used
with gerrit before and we download the hook, then we do want to amend the
commit.

Change-Id: I29d3ec42ca086ecdfec6dd788f04e19649eacb24