Due to gerrit bug regeneration of Change-Id must be nessecary
Simultaneously dirrect push to master and uploading a gerrit
patch for review to some branches is currently not supported.
To circumvent it a new option --new-changeid is provided.

git-review --dry-run also supported now: only the command
is written without actually execute it.

Special thank goes to Roan Kattouw for helping out with
GIT_EDITOR command.

Bug #1032675

Change-Id: Id1433c11ce9f0a2ad51b3bdfa581fdb86cff8f0e