Get rid of "status" parameter in run_command
This is the first commit in a refactoring attempt
of the SSH/SCP connection code that got abandoned
as https://review.openstack.org/#/c/5720/.

Let's get rid of "status" parameter, effectively
making "run_command" just ignore the return code.

This is needed to introduce (*argv, **envp)
as more pythonic parameters to "run_command".

Change-Id: If6b56e567a04170c406c9f686c7e1aa9cd4230c1