Introduce CommandFailed exception for run_command
This exception should print the error message
as well as the offending command with the parameters
and its output.

This way lots of status checking code (if (status != 0))
can be removed in the future.

An example CannotInstallHook exception has been introduced.

Example output in case of failure:

Problems encountered installing commit-msg hook
The following command failed with exit code 1
    "scp -P 29418 saper@blabla:hooks/commit-msg .git/hooks/commit-msg"
-----------------------
ssh: Could not resolve hostname blabla: Name or service not known
-----------------------

This way only one place will be used to control output of errors,
it is also simple to introduce separate exit code for every error
in git-review.

Change-Id: I92b8637c87df3bdc328faef456a8391305fb92cb