Review list fields should have constant widths
When asking for a list of review, the decimal change number can have a
vary length which could cause the list to not be properly aligned.

This patch keeps track of the widths of particular fields,
applies color control characters and crafs a format string.

Change# and branch are right justified, commit entry is
left justified.

Change-Id: I2d9114ed76a40cd029e96b69f5e1c2df9ff430f9