Topic name should be determined only for git push
When setting up git-review, it tries to acquire a topic name although
that is of no use to setup.  A side effect is that git-review can not be
setup whenever the git remote is a fresh repository since it get no
branch. Solve this by moving it after the setup routines.

Inform the user why the particular topic name has been selected.

Change-Id: Iaa942b8dc77b0c3f720085358c38e8295138025a