Better determine git directories
- Call "git-rev-parse" only once, not twice
- In case where current directory is unreadable,
the current message is:

No '.gitreview' file found in this repository.
We don't know where your gerrit is. Please manually create
a remote named gerrit and try again.

(which is not true, because .gitreview *is* present)

After this change it fails a bit better:

Cannot determine where .git directory is
The following command failed with exit code 128
    "git rev-parse --show-toplevel --git-dir"
-----------------------
fatal: Unable to read current working directory: No such
file or directory
-----------------------

Change-Id: Idbb1e68b689f226d9434db131db5d9765752a315