Don't fetch remote outside of rebase flow.
The assert_one_change code checks your changes against the set of changes in
the branch that you're submitting to. It's not necessary to have the absolute
latest changes to do the safety check ... the worst that will happen is that
you'll be prompted about a patch that's already merged.

However, if you run without -R, you'll get the update fetched anyway.

If you run _WITH_ -R, then we're doing a network access that you don't need
anyway.

Change-Id: I466aa81a521f784d9afc1dbae914c68a33ff79f5