Provide usage help even if not in Git directory
If a new user installs git-review and invokes it with no arguments (or with
'-h' / '--help') outside of a Git working tree, all she gets is a terse error
message that offers no usage help.

With this change, git-review behaves in the following manner when invoked
outside a git working tree:

* If a command-line argument is present and it is not '-h' or '--help', it is
  assumed that the user was attempting some action, and so git-review fails
  with the same error message as before.
* If no command-line arguments are present, or if there is a single
  command-line argument that is either '-h' or '--help', usage information is
  printed and the program exits.

This patch also adds myself to AUTHORS and fixes a small typo in README.rst.

Change-Id: I6fda72bf5311e74318b42e2860e7742e07b515de