Expand multiple commit submission warning.
Added some more information to the message given to users when they
attempt to submit a review with multiple commits. This reassures users
who are submitting dependent changes that multiple commit submission is
expected in that case.

Also moves the message about potentialy squashing to before the prompt
and confirms that the submission will be cancelled if not confirmed.

Change-Id: I65d5a73c27d0be0170957598da00a69d1c44a8b6