Fix pep8 deps for pyflakes version conflict.
Fixes bug 1199400.

* tox.ini: In the pep8 environment, flake8 and hacking were both
listed as deps. This was fine while both worked with the latest
pyflakes, but then pyflakes 0.7.3 came along and was getting
installed before the more explicit pyflakes 0.7.2 requirement was
parsed for hacking. Since flake8 is in the requirements list for
hacking anyway, this just removes it from the pep8 environment deps
list.

Change-Id: I647235a4f6d109e00f13a156c57ff3a0e9a13a81