Preserve merges when doing a rebase
When a person is using git-review after having merged in a branch, the
default rebase behavior as currently implemented will rebase all of the
commits that are inside of the merge. This is almost never what someone
wants to do.

Add -p to the rebase commands which will tell git to preserve the merges
and only rebase the merge commit itself.

Explicitly preserve original SHA1 from before the rebase, since git may
not set ORIG_HEAD as expected depending the operations performed.

Change-Id: I8a5f0307113cd0cf0d08c664c13a0676400bd05c