Ignore content of merge commits in reporting
When we report changes that are going to be submitted, the results of
merging a giant branch into the current one is TERRIFYING, as it shows
every commit in the branch.

Exclude all changes that are already on branches in the remote so
that only the merge commit and its first parent are shown, which
should make the output less crazy-inducing.

Change-Id: I461f3eb23642deaf2caded631291414ae6997a91