Provide nicer user message for missing remote ref
If the remote branch being targetted does not exist, any attempted
rebase will fail with an error that is largely unhelpful for users.

Assuming the preceding fetch succeeded properly, the most common reason
for the remote reference to not exist, is that the user wants the
changes to go to a new branch rather than the default. In this case
provide the user with a more helpful message indicating that if this is
the intention, to run git-review with the option to disable the rebase
in order to get the desired outcome of gerrit creating the branch
desired for the new reviews.

Change-Id: I4c92e901950abf2c879d00660daae3ccce7ddfc8