Use 'no_proxy' env variable in addition to uppercase
Need to override both variations of the no_proxy environment variable
since some tools do not check for the other if one is already defined.

While 'no_proxy' appears to be now the most common convention, it is
safer to set both lower and upper case to the same value when running
unit tests.

Change-Id: I536b6f2711ac70e0a2a45676098fb461ab3b8d33