Add a note about contribution priorities
It's been poorly documented so far that the current git-review
maintainers are avoiding addition of new features. Gerrit-based
change workflows are already confusing enough to new users, and
providing them with too many complex options risks making things
worse. This tool's goal is to do a few things well and simply in
support of an opinionated workflow, without becoming everything to
everyone.

Since git-review sees broad usage outside the OpenStack contributor
community, it is subject to a lot of drive-by feature additions. Add
a note to the contributing document indicating its maintainers
consider it to be mostly complete (modulo bug fixes, compatibility
improvements, adding test coverage), so that readers will consider
strongly whether their particular improvement is well suited to the
existing scope of the tool.

Change-Id: I11b9f898286b75177493af31e209c2853d78f245