Fix no_git_dir UnboundLocalError in except block
In Python 3, exception blocks clean their variables up, and we get
a UnbloundLocalError when we try to reference the target outside
of the except block.

To pass the exception out we set a variable from the target.

Also pull --license above raising the error, as it should work
outside a git repo too.

https://www.python.org/dev/peps/pep-3110/#semantic-changes

Change-Id: I298ac3648068cd07c4cf232e32ab3daaaf8e549d