Fix compare_review's use of fetch_review
fetch_review was updated to take project config information so that
projects hosted over http below some root (eg
https://gerrit.org/gerrit/here) could be properly fetched against. But
in adding this we broke the compare_review functionality which is the
other user of fetch_review.

Fix this by passing project config info into compare_review so that it
can pass that along to fetch_review.

Additionally we need to compare like types when looking at patchset
numbers so coerce them to ints when comparing.

Change-Id: I98fbf31821dc3a9162700725dec07bc7685ea5ca