`work-in-progress` and `private` workflow in Gerrit 2.15
In Gerrit 2.15 two new workflow states(`work-in-progress` and
`private`) are introduced. This patch tries to implement them.

Gerrit documentation says:

To push a private change or to turn a change private on push the private
 option can be specified:

  git push ssh://john.doe@git.example.com:29418/kernel/common HEAD:refs/for/master%private

Omitting the private option when pushing updates to a private change
doesn’t make change non-private again. To remove the private flag from a
 change on push, explicitly specify the remove-private option:

  git push ssh://john.doe@git.example.com:29418/kernel/common HEAD:refs/for/master%remove-private

To push a wip change or to turn a change to wip the work-in-progress
(or wip) option can be specified:

  git push ssh://john.doe@git.example.com:29418/kernel/common HEAD:refs/for/master%wip

Omitting the wip option when pushing updates to a wip change doesn’t
make change ready again. To remove the wip flag from a change on push,
 explicitly specify the ready option:

  git push ssh://john.doe@git.example.com:29418/kernel/common HEAD:refs/for/master%ready

https://gerrit-documentation.storage.googleapis.com/Documentation/2.15/intro-user.html#private-changes
https://gerrit-documentation.storage.googleapis.com/Documentation/2.15/intro-user.html#wip

Change-Id: Ia093e4a691fa8eb17c798473a79a97694202fd03