Use remote_url instead of remote for download
If usepushurl is set, then it's likely that the pushurl is pointing
to the code review system while the normal url is pointing to a read
only mirror. If the read-only mirror in question doesn't contain
the gerrit refs (like opendev.org currently) then git fetch origin
won't work properly. However, we already calculate the correct
URL at the top of the function for use in query_reviews, and fetch
will happily work with a full url rather than a named remote.

Update the function to use the remote_url so that usepushurl works
even if the mirror url does not contain refs/changes.

Change-Id: Ib72afe97e65cb1dcaf95e28450dfe6e7d5f88965